backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.8.0
pydantic-settings>=2.0.0
python-dotenv==1.0.0
httpx==0.25.2
openai==1.3.7
anthropic==0.7.8
python-multipart==0.0.6
jinja2==3.1.2
loguru==0.7.2
pyinstaller==6.15.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
asyncio-mqtt